+++ resources/views/assets/edit.blade.php
@@ -36,7 +36,7 @@
                 </div>
 
             {!!Form::close()!!}
-            {!! Form::open(['method'=>'DELETE', 'action'=> ['AssetsController@destroy', $asset->id]]) !!}
+            {!! Form::open(['method'=>'DELETE', 'action'=> ['AssetsController@destroy', $asset->id],'onsubmit'=>"return confirm('Are you sure you want to delete this item?')"]) !!}
 
             <div class="form-group">
                 {!! Form::submit('Delete Asset', ['class'=>'btn btn-danger col-sm-6']) !!}
